Change QuatityValue value type to BigDecimal

+++ shetland/src/main/java/org/n52/shetland/inspire/omso/ProfileObservation.java
@@ -91,7 +91,7 @@ public class ProfileObservation
                 }
                 if (level.isSetLocation()) {
                     Coordinate coordinate = level.getLocation().getCoordinate();
-                    coordinate.z = level.getLevelStart().getValue();
+                    coordinate.z = level.getLevelStart().getValue().doubleValue();
                     coordinates.add(coordinate);
                     if (srid == 0) {
                         srid = level.getLocation().getSRID();
